requirements.txt
einops==0.8.0
nlg_metricverse==0.9.9
numpy==1.26.4
ordered_set==4.1.0
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.3.0
stanza==1.6.1
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.1
torchvision==0.16.1
tqdm==4.65.0
transformers==4.36.2
translators==5.9.2
